Add an "unprotectedPorts" exception mechanism
For Telco specifically, there is a need to allow PTP traffic even if a
port is unathenticated.

This change introduces a generic mechanism that can be used to configure
an arbitrary set of ingress TCP and/or UDP ports that will be allowed
even if an 802.1x-protected port is unauthenticated.  Use with caution.

Signed-off-by: Jim Ramsay <jramsay@redhat.com>