Implement VLAN based VF traffic control mechanism
This configures associated VFs for a given unauthenticated PF with reserved VLAN
value 4095 and set tx queue to disable state. As soon as first wpa client gets
authenticated, it restores original VLAN value and make tx queue state to auto.
At the same time it also keeps track of VLAN update on the interface and
configure the VF according to PF's authenticated state.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>